+++ n0core/scheduler/agent/volumer.py
@@ -1,4 +1,4 @@
-from client import consumer, logger, send  # NOQA
+from scheduler import consumer, logger, send  # NOQA
 try:
     from n0core.lib import proto
 except:  # NOQA
